Install.base now supports passing the force parameter
Fixed XML namespace in several nuspec files
Building the dev.extension now removes the modules if it it's installed
The install choco script now upgrades Chocolatey-Personal instead of installing it
Updated package.base to build all the dependencies

+++ Scripts/build.chocolatey-dev.extension.ps1
@@ -1,13 +1,16 @@
 $packagesDir = Join-Path $PSScriptRoot '..\..\..\BoxStarter' -Resolve
 $parentDir = Join-Path $PSScriptRoot '..\' -Resolve
 
-choco uninstall Chocolatey-Dev.extension -f -ErrorAction SilentlyContinue
-Remove-Item -Recurse C:\ProgramData\Chocolatey\extensions\chocolatey-dev -ErrorAction SilentlyContinue
+if ((choco list Chocolatey-Dev -lo -r)) {
+    choco uninstall Chocolatey-Dev.extension -f
+    Remove-Item -Recurse C:\ProgramData\Chocolatey\extensions\chocolatey-dev -ErrorAction SilentlyContinue
+}
 
 & $parentDir\build-push.ps1 NuGet-Commandline -f
 & $parentDir\build-push.ps1 Octopus-Tools -f
-#& $parentDir\build-push.ps1 psake
-& $parentDir\build-push.ps1 pscx -f
+& $parentDir\build-push.ps1 Powershell-Carbon -f
+& $parentDir\build-push.ps1 Powershell-Psake -f
+& $parentDir\build-push.ps1 Powershell-CommunityExtensions -f
 & $parentDir\build-push.ps1 Octopus-Tools -f
 & $parentDir\build-push.ps1 Chocolatey-Dev.extension -f
